refactor(api): déplacer les contrôleurs createComment et moderateComment vers shantytownComment refers to https://trello.com/c/evsOelCU/2717
- déplace createComment et moderateComment de controllers/shantytown vers controllers/shantytownComment
- renomme les fichiers en shantytownComment.* pour cohérence
- nomme les fonctions de route (createCommentRoute, moderateCommentRoute) et de contrôleur (createComment, moderateComment)
- met à jour l'import du spec createComment
