+++ packages/api/server/controllers/shantytownComment/createComment/shantytownComment.createComment.route.ts
@@ -1,10 +1,10 @@
 import { type ApplicationWithCustomRoutes } from '#server/loaders/customRouteMethodsLoader';
-import validator from './shantytown.createComment.validator';
-import controller from './shantytown.createComment';
+import validator from './shantytownComment.createComment.validator';
+import controller from './shantytownComment.createComment';
 
-export default (app: ApplicationWithCustomRoutes): void => {
+export default function createCommentRoute(app: ApplicationWithCustomRoutes): void {
     app.customRoutes.post('/towns/:id/comments', controller, validator, {
         authenticate: true,
         multipart: true,
     });
-};
+}
