+++ packages/api/server/controllers/shantytownComment/createComment/shantytownComment.createComment.spec.ts
@@ -7,7 +7,7 @@ import { serialized as fakeUser } from '#test/utils/user';
 import shantytownCommentService from '#server/services/shantytownComment';
 import ServiceError from '#server/errors/ServiceError';
 import { NextFunction } from 'express';
-import ctlCreate from './shantytown.createComment';
+import ctlCreate from './shantytownComment.createComment';
 
 
 const { expect } = chai;
