+++ packages/api/server/controllers/shantytownComment/moderateComment/shantytownComment.moderateComment.route.ts
@@ -1,9 +1,9 @@
 import { type ApplicationWithCustomRoutes } from '#server/loaders/customRouteMethodsLoader';
-import controller from './shantytown.moderateComment';
+import controller from './shantytownComment.moderateComment';
 
-export default (app: ApplicationWithCustomRoutes): void => {
+export default function moderateCommentRoute(app: ApplicationWithCustomRoutes): void {
     app.customRoutes.delete('/towns/:id/comments/:commentId', controller, undefined, {
         authenticate: true,
         multipart: false,
     });
-};
+}
